fixed missing time

+++ Core/Model/MissingTime/Time.cs
@@ -5,7 +5,7 @@
         public Guid UserId{ get; set; }
         public string UserName { get; set; }
         public int TimekeeperIndex { get; set; }
-        public decimal WorkHrs { get; set; }
+        public int WorkHrs { get; set; }
         public DateTime WorkDate { get; set; }
         public decimal WeeklyHoursRequired { get; set; }
         public string Workcalendar { get; set; }
